+++ qmlbind/src/exporter.cpp
@@ -27,7 +27,11 @@ Exporter::Exporter(const char *className, qmlbind_client_class *classObject, qml
     mBuilder.setClassName(className);
 }
 
-static QByteArray methodSignature(const char *name, int arity)
+std::unique_ptr<QMetaObject, decltype(&free)> Exporter::toMetaObject() const {
+    return std::unique_ptr<QMetaObject, decltype(&free)>(mBuilder.toMetaObject(), free);
+}
+
+QByteArray Exporter::methodSignature(const char *name, int arity)
 {
     QByteArray sig;
     sig += name;
@@ -46,7 +50,7 @@ static QByteArray methodSignature(const char *name, int arity)
 
 void Exporter::addMethod(const char *name, int arity)
 {
-    QMetaMethodBuilder method = mBuilder.addMethod(methodSignature(name, arity), "QJSValue");
+    QMetaMethodBuilder method = mBuilder.addMethod(Exporter::methodSignature(name, arity), "QJSValue");
 
     Method methodInfo;
     methodInfo.arity = arity;
@@ -57,7 +61,7 @@ void Exporter::addMethod(const char *name, int arity)
 
 void Exporter::addSignal(const char *name, const QList<QByteArray> &args)
 {
-    QMetaMethodBuilder method = mBuilder.addSignal(methodSignature(name, args.size()));
+    QMetaMethodBuilder method = mBuilder.addSignal(Exporter::methodSignature(name, args.size()));
     method.setParameterNames(args);
 
     mSignalIndexMap[name] = method.index();