+++ qmlbind/src/metaobject.h
@@ -13,18 +13,21 @@ class SignalEmitter;
 class MetaObject : public QMetaObject, public std::enable_shared_from_this<MetaObject>
 {
 public:
-    MetaObject(const std::shared_ptr<const Exporter> &exporter);
-    ~MetaObject();
+    MetaObject(std::unique_ptr<const Exporter> &&exporter, QMetaObject* prototype);
+    MetaObject(std::unique_ptr<const Exporter> &&exporter, std::unique_ptr<QMetaObject, decltype(&free)> prototype);
 
     Wrapper *newWrapper(qmlbind_client_object *object) const;
     Wrapper *newObject(void *memory) const;
+
+    // needed because indexOfSignal() needs the full normalized signal signature, not only the name.
+    int indexOfSignalName(const char *name) const;
 
     int metaCall(QObject *object, Call call, int index, void **argv) const;
 
 private:
 
-    std::shared_ptr<const Exporter> mExporter;
-    std::shared_ptr<QMetaObject> mPrototype;
+    std::unique_ptr<const Exporter> mExporter;
+    std::unique_ptr<QMetaObject, decltype(&free)> mPrototype;
 };
 
 } // namespace QmlBind